# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a Social Network - OnlyFans api specifications API REST .",
    "version": "1.0.2",
    "title": "Swagger Api-profile",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "yannick.simon@kysoe.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/profile": {
      "post": {
        "tags": [
          "profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "profile post",
        "description": "profile post desccription",
        "operationId": "profilePost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Profile"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "profile list",
        "description": "profile desccription",
        "operationId": "profileList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/profile/{profile_id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "profile get",
        "description": "profile get desccription",
        "operationId": "profileGet",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "profile delete",
        "description": "profile delete desccription",
        "operationId": "profileDelete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
...
      "Profile": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        },
        "description": "Profile object that needs to be add",
        "required": true
      },
...
    },
    "schemas": {
...
      "Profile": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Profile"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}